pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "srt-voiceover"
version = "0.1.0"
description = "Convert SRT subtitle files to synchronized voiceover audio using Edge TTS"
readme = "README.md"
authors = [
    {name = "Nathan Jones", email = ""},
]
license = {text = "MIT"}
requires-python = ">=3.7"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Multimedia :: Video",
]
keywords = ["srt", "subtitle", "voiceover", "tts", "edge-tts", "dubbing", "audio"]

dependencies = [
    "pysrt>=1.1.2",
    "pydub>=0.25.1",
    "requests>=2.31.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=22.0",
    "flake8>=5.0",
    "mypy>=0.990",
]

[project.urls]
Homepage = "https://github.com/leakydata/srt-voiceover"
Repository = "https://github.com/leakydata/srt-voiceover"
Issues = "https://github.com/leakydata/srt-voiceover/issues"

[project.scripts]
srt-voiceover = "srt_voiceover.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py37', 'py38', 'py39', 'py310', 'py311']

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

